+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Not released
 ------------
 
+* Update dependancies and add python 3.10 to CI (#49)
+
 v0.4.2 (2022/05/31)
 -------------------