+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Not released
 * Update dependancies and add python 3.10 to CI (#49)
 * Optimize bitwise AND in SASCA: O(nc*log(nc)) instead of O(nc^2).
 * SASCA: add `SASCAGraph.sanity_check`.
+* SASCA: allow tables to not be bijective.
 
 v0.4.2 (2022/05/31)
 -------------------