+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 Not released
 ------------
 
+v0.4.3 (2022/10/27)
+-------------------
+
+* Upgrade CI to 3.11.
 * Update dependancies and add python 3.10 to CI (#49)
 
 v0.4.2 (2022/05/31)